pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chronoapi"
version = "1.0.0"
description = "TTM time-series forecasting API (FastAPI)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115.6,<0.116",
    "uvicorn[standard]>=0.32.1,<0.33",
    "gunicorn>=23.0.0,<24",
    "pandas>=2.2.3,<3",
    "torch>=2.5.1,<3",
    "transformers>=4.46.0,<5",
    "plotly>=5.24.0,<6",
    "granite-tsfm @ git+https://github.com/ibm-granite/granite-tsfm.git",
    "aiokafka>=0.12.0,<0.13",
    "prometheus-client>=0.21.0,<0.22",
    "pydantic>=2.10.0,<3",
    "pydantic-settings>=2.6.0,<3",
]

[project.urls]
Repository = "https://github.com/fraware/chronoapi"

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0,<0.29",
    "pytest>=8.3.0,<9",
    "pytest-asyncio>=0.24.0,<0.25",
    "pytest-cov>=6.0.0,<7",
    "ruff>=0.8.0,<0.9",
    "mypy>=1.13.0,<2",
]
otel = [
    "opentelemetry-api>=1.28.0,<2",
    "opentelemetry-sdk>=1.28.0,<2",
    "opentelemetry-instrumentation-fastapi>=0.49b0,<0.50",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-q --cov=app --cov-report=term-missing --cov-fail-under=25"

[tool.coverage.run]
branch = true
source = ["app"]
omit = ["app/__init__.py"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true
packages = ["app"]
